requirements.txt
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
absl-py==2.1.0
python-dotenv==1.0.1
lightning==2.4.0